+++ apps/main-app/ts/features/itwallet/navigation/saga/index.ts
@@ -4,24 +4,23 @@ import { call } from "typed-redux-saga/macro";
 import type { ItwDeepLink } from "../utils";
 
 import NavigationService from "../../../../navigation/NavigationService";
-import { itwLinkingConfig } from "../../navigation/useItwLinkingOptions";
+import { itwLinkingConfig } from "../useItwLinkingOptions";
 
 /**
  * Converts a parsed ITW deep link into a navigation action and dispatches it.
  * The caller owns navigator readiness and stored-link clearing.
  */
 export function* handleItwStoredDeepLink(deepLink: ItwDeepLink) {
-  const itwState = getStateFromPath(deepLink.path, itwLinkingConfig);
-  const itwAction =
-    itwState === undefined
+  const state = getStateFromPath(deepLink.path, itwLinkingConfig);
+  const action =
+    state === undefined
       ? undefined
-      : getActionFromState(itwState, itwLinkingConfig);
+      : getActionFromState(state, itwLinkingConfig);
 
-  if (itwAction === undefined) {
+  if (action === undefined) {
     return false;
   }
 
-  yield* call(NavigationService.dispatchNavigationAction, itwAction);
-
+  yield* call(NavigationService.dispatchNavigationAction, action);
   return true;
 }
