+++ apps/main-app/ts/features/linking/sagas/index.ts
@@ -6,8 +6,8 @@ import {
   shouldTriggerWalletUpdate
 } from "../../../utils/deepLinkUtils";
 import { cgnEycaStatus } from "../../bonus/cgn/store/actions/eyca/details";
-import { handleItwStoredDeepLink } from "../../itwallet/linking/saga";
-import { parseItwDeepLink } from "../../itwallet/linking/utils";
+import { handleItwStoredDeepLink } from "../../itwallet/navigation/saga";
+import { parseItwDeepLink } from "../../itwallet/navigation/utils";
 import { initiateAarFlow } from "../../pn/aar/store/actions";
 import { isSendAarLink } from "../../pn/aar/utils/deepLinking";
 import { walletUpdate } from "../../wallet/store/actions";
@@ -27,16 +27,11 @@ export function* handleStoredLinkingUrlIfNeeded() {
 
     const itwDeepLink = parseItwDeepLink(storedLinkingUrl);
     if (itwDeepLink !== undefined) {
+      yield* put(clearLinkingUrl());
       yield* call(waitForMainNavigator);
-      const didHandleItwDeepLink = yield* call(
-        handleItwStoredDeepLink,
-        itwDeepLink
-      );
+      yield* call(handleItwStoredDeepLink, itwDeepLink);
 
-      if (didHandleItwDeepLink) {
-        yield* put(clearLinkingUrl());
-        return true;
-      }
+      return true;
     }
 
     if (shouldTriggerWalletUpdate(storedLinkingUrl)) {
